label generic capabilities
Some Addon requirements are generic: storageclass, cni, ingresscontroller, etc.
These could be replaced by other sets of software that provide the same
functionality so we don't want to configure dependencies to require
specific software packages, but rather a generic capability.

This patch adds labels that can be matched for a generic capability.